+++ src/parser/mrt/messages/mod.rs
@@ -3,6 +3,7 @@ use crate::models::{AsnLength, Bgp4MpEnum, Bgp4MpType, MrtMessage, TableDumpV2Me
 use bytes::Bytes;
 
 pub(crate) mod bgp4mp;
+pub(crate) mod legacy_bgp;
 pub(crate) mod table_dump;
 pub(crate) mod table_dump_v2;
 
@@ -10,6 +11,11 @@ impl MrtMessage {
     pub fn encode(&self, sub_type: u16) -> Result<Bytes, EncodingError> {
         let msg_bytes: Bytes = match self {
             MrtMessage::TableDumpMessage(m) => m.encode()?,
+            MrtMessage::TableDumpMessageBatch(messages) => {
+                crate::parser::mrt::messages::table_dump::encode_table_dump_batch(
+                    messages, sub_type,
+                )?
+            }
             MrtMessage::TableDumpV2Message(m) => match m {
                 TableDumpV2Message::PeerIndexTable(p) => p.encode()?,
                 TableDumpV2Message::RibAfi(r) => r.encode()?,
@@ -46,6 +52,9 @@ impl MrtMessage {
                         msg.encode(asn_len)?
                     }
                 }
+            }
+            MrtMessage::LegacyBgp(m) => {
+                crate::parser::mrt::messages::legacy_bgp::encode_legacy_bgp(m, sub_type)?
             }
         };